+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- 
+- In the generated config **dependency** `truehd` should be `truehdd`. This is fixed when you generate a new config, but you can manually modify it.
 
 ### Removed