+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.4] - 2025-09-14
+
+### Fixed
+
+- Atmos streaming mode was missing 1024 bitrate selection.
+- Atmos bluray mode was defaulting to 448, should have been 1280 (this would have been corrected automatically).
+
 ## [1.2.3] - 2025-09-14
 
 ### Added